refactor: move home features to app/[locale]/components and update all importers
