+++ app/app/[locale]/components/carousel/utils.ts
