+++ app/app/[locale]/components/channel_shop/index.tsx
