+++ app/app/[locale]/components/restrict_search/index.tsx
