+++ app/app/[locale]/components/restrict_search/types/index.ts
