+++ app/features/category/index.tsx
@@ -5,7 +5,7 @@ import { useParams } from 'next/navigation';
 import { RecommandCategories } from 'app/[locale]/[title]/components/category/recommand-categories';
 import Title from './title';
 import { Shop } from 'components';
-import RestrictSearch from 'features/restrict_search';
+import RestrictSearch from 'app/[locale]/components/restrict_search';
 import {
   categoryByTitleQueryOptions,
 } from '../../app/[locale]/[title]/queries';
