+++ app/features/index.ts
@@ -1,17 +1,4 @@
-import Carousel from './carousel';
-import Layout from './layout';
-import Shortcut from './shortcut';
-import RestrictSearch from './restrict_search';
-import Channel from './channel';
-import Category from './category';
-import Store from './store';
+import Category from "./category";
+import Store from "./store";
 
-export {
-	Carousel,
-	Layout,
-	Shortcut,
-	RestrictSearch,
-	Channel,
-	Category,
-	Store
-};
+export { Category, Store };
